# src/main/resources/OpenApi.yaml
openapi: 3.0.3
info:
  title: API de proyecto web para reserva de alojamientos
  description: |-
    Api rest para gestion de alojamientos, usuarios y reservas.
    
    El proyecto es desarrola por la parte de Bakend en spring boot y uso angular para el fronted,
      además del uso de Maria DB como base de datos relacional.
    
    Enlaces de utilidad:
      -[Enlace de GitHub](https://github.com/Elkinlol/Proyecto-final)
      
    Contacto: elkin.bermudezg@uqvirtual.edu.co
  version: 1.1
servers:
  - url: http/:localhost:8082/api
    description: servidor local
...
paths:
...
  /users/{id}/password:
    patch:
      tags: [Users]
      summary: Actualizar contraseña de un usuario
      operationId: ActualizarContrasenia
      security:
        - BearerAuth: [ ]
      parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updatePassword'
            example:
              currentPassword: "ContraseñaVieja123"
              newPassword: "NuevaContraseñaSegura456"
      responses:
        '200':
          description: contraseña actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                response: 'la contraseña ha sido actualizada'
        '400':
          description: Faltan datos requeridos o contraseña debil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'la contraseña debe tener minimo 8 caracteres, incluyendo numeros, letras y caracteres especiales'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'la contraseña actual no coincide'
        '403':
          description: Prohibido -- Sin permisos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'No tiene permiso para la actualizacion de la contraseña'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'No se encuentra el usuario'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'error inesperado, intentar más tarde'
        '503':
          description: Sin disponibilidad de servicio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Error inesperado, intente más tarde"
  /users/{id}:
    put:
      tags : [Users]
      summary: actualizar perfil(nombre, foto, telefono)
      operationId: UpdateProfile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
          type: String
          description: Id del usuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              nombre: "Carlos Ramírez"
              telefono: "+57 3001234567"
              fotoPerfil: "https://miservidor.com/perfiles/carlos.jpg"
              descripcion: "Anfitrión con experiencia en Medellín"
      responses:
        '200':
          description: Perfil actualizado con exito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        "400":
          description: Datos inválidos (formato incorrecto o campos vacíos)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Número de teléfono inválido"
        "401":
          description: Usuario no autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "No autorizado, inicie sesión"
        "403":
          description: Intento de editar el perfil de otro usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "No tiene permisos para editar este perfil"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Error inesperado, intente más tarde"
    get:
      tags: [ Users ]
      description: Ver el perfil de un usuario
      operationId: ViewProfile
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
          type: Integer
          description: Id del usuario
      responses:
        '200':
          description: Ver perfil concretado
          content:
            schema:
              $ref: "#/components/schemas/viewUsers"
        '401':
          description: Usuario NO autorizado
          content:
            schema:
              $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: No tiene  los privilegios necesarios
          content:
            schema:
              $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Usuario no encontrado
          content:
            schema:
              $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error del sevidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Se cambia el estado de un usuario
      tags: [Users]
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: 1134
            type: string
      responses:
        "200":
          description: Usuario eliminado con exito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/deleteUserRespose"
        "401":
          description: No tiene permisos necesarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: No tiene acceso a ese recurso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /listings/{id}:
    put:
      summary: Editar un alojamiento existente
      tags:  [Listings]
      operationId: editListing
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del alojamiento a editar
          schema:
            type: integer
            example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Alojamiento actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: El usuario no es propietario del alojamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Eliminar un alojamiento (soft delete)
      tags: [Listings]
      operationId: deleteListing
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del alojamiento a eliminar
          schema:
            type: integer
            example: 101
      responses:
        "200":
          description: Alojamiento marcado como eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteListingResponse"
        "400":
          description: El alojamiento no puede eliminarse (reservas futuras)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: El usuario no es propietario del alojamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Listings]
      summary: Obtener los alojamientos -- ver sus detalles
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del alojamiento a eliminar
          schema:
            type: integer
            example: 101
      responses:
        '200':
          description: Se obtiene el alojamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        '401':
          description: No tiene autorizacion
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: No permisos para realizar esta opcion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
                
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reserves:
    post:
      summary: Creacion de una reserva
      tags: [Reservations]
      operationId: MakeReserv
      security:
        - bearerAuth: [ ]
      bodyRequest:
        content:
          schema:
            $ref: "#/components/schemas/MakeReservRequest"
      responses:
        '201':
          description: Reserva hecha exitosamente
          content:
            schema:
              $ref: "#/components/schemas/ReservResponse"
        '400':
          description: validaciones fallidas
          content:
            schema:
              $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Alojamiento no encontrado
          content:
            schema:
              $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: fechas de reserva no disponibles
          content:
            schema:
              $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: error de servidor
          content:
            schema:
              $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [ Reservations ]
      summary: Listar reservas del usuario autenticado
      operationId: listUserReservations
      security:
        - bearerAuth: [ ]
      parameters:
        - name: estado
          in: query
          required: false
          description: Filtrar reservas por estado (Pendiente, Confirmada, Cancelada, Completada)
          schema:
            type: string
            enum: [ Pendiente, Confirmada, Cancelada, Completada ]
        - name: fechaInicio
          in: query
          required: false
          description: Filtrar reservas a partir de una fecha de inicio
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          required: false
          description: Filtrar reservas hasta una fecha de fin
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          description: Número de página para la paginación (por defecto 1)
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          description: Tamaño de página para la paginación (por defecto 10)
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Lista de reservas del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 25
                  page:
                    type: integer
                    example: 1
                  size:
                    type: integer
                    example: 10
                  reservas:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReservResponse"
        "401":
          description: No autorizado (JWT inválido o ausente)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reserves/{id}:
    get:
      tags: [ Reservations ]
      summary: Ver detalles de una reserva
      operationId: getReservationById
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva
          schema:
            type: integer
      responses:
        "200":
          description: Detalles de la reserva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservResponse"
        "401":
          description: No autorizado (JWT inválido o ausente)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reserva no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Cancelar una reserva
      tags: [Reservations]
      operationId: cancelReserv
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva a cancelar
          schema:
            type: string
            example: 2341
      responses:
        '200':
          description: Reserva cancelada de manera exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        '400':
          description: No se puede cancelar la reserva (El tiempo no es adecuado)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: No se encuentra autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: No tiene permisos para realizar esta acción
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Reserva no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '503':
          description: Servicio no disponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    deleteUserRespose:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "El Usuario con id 123 fue eliminado"
...
    ReservResponse:
      type: object
      properties:
        id:
          type: integer
          example: 101
        alojamientoId:
          type: integer
          example: 123
        usuarioId:
          type: integer
          example: 45
        fechaInicio:
          type: string

          example: "2025-09-10"
        fechaFin:
          type: string

          example: "2025-09-15"
        numeroHuespedes:
          type: integer
          example: 4
        estado:
          type: string
          enum: [ Pendiente, Confirmada, Cancelada, Completada ]
          example: Confirmada
        fechaCreacion:
          type: string
          format: date-time
          example: "2025-09-07T14:23:00Z"
        fechaActualizacion:
          type: string
          format: date-time
          example: "2025-09-07T14:25:00Z"
      required:
        - id
        - alojamientoId
        - usuarioId
        - fechaInicio
        - fechaFin
        - numeroHuespedes
        - estado
    MakeReservRequest:
      type: object
      required:
        - alojamientoId
        - fechaIngreso
        - fechaSalida
        - cantHuespedes
      properties:
        alojamientoId:
          type: integer
          example: 123
        fechaIngreso:
          type: string
          format: date
          example: 2026-06-04
        fechaSalida:
          type: string
          format: date
          example: 2026-06-07
        cantHuespedes:
          type: integer
          example: 2
...
    ListingResponse:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ubicacion:
          type: object
        PriceForNight:
          type: number
        Capacity:
          type: integer
        imagenes:
          type: array
          items:
            type: string
            format: uri
        anfitrionId:
          type: integer
          example: 55
        fechaPublicacion:
          type: string
          format: date-time
          example: "2025-09-07T13:00:00Z"
      required:
        - title
        - ubicacion
        - priceForNight
        - description
        - capacity
    DeleteListingResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "El alojamiento con id 123 fue eliminado"
    UpdateListingRequest:
      type: object
      properties:
        title:
          type: string
          example: Pent house
        descition:
          type: string
          example: Las mejores vistas de medellin
        ubicacion:
          type: object
          example: {direccion: calle 14, ciudad: medellin, latitud: 100, longitud: 200}
        priceForNight:
          type: number
          format: float
          example: 200
        Capacity:
          type: integer
          example: 40
        imagenes:
          type: array
          items:
            type: string
            format: uri
...
    viewUsers:
      Usuario:
        type: object
        properties:
          id:
            type: integer
            example: 12
          nombre:
            type: string
            example: "Carlos Pérez"
          email:
            type: string
            format: email
            example: "carlos.perez@mail.com"
          telefono:
            type: string
            example: "+57 3123456789"
          rol:
            type: string
            enum: [ usuario, anfitrion ]
            example: "usuario"
          fechaNacimiento:
            type: string
            format: date
            example: "1995-06-20"
          fotoPerfil:
            type: string
            format: uri
            nullable: true
            example: "https://cdn.miapp.com/usuarios/12/perfil.jpg"
          descripcion:
            type: string
            nullable: true
            example: "Anfitrión con experiencia en Medellín."
...
    UpdateProfileRequest:
      type: object
      properties:
        rol:
          type: Rol
          description: Rol del usuario actual
          enum: [Host, Guest]
        nombre:
          type: string
          maxLength: 100
        telefono:
          type: string
          pattern: "^[+0-9]{7,15}$"
        fotoPerfil:
          type: string
          format: uri
        descripcion:
          type: string
          maxLength: 500
          description: Solo visible para anfitriones
      description: Campos opcionales para actualizar el perfil
    UpdateProfileResponse:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        email:
          type: string
          format: email
        telefono:
          type: string
        rol:
          type: string
          enum: [ usuario, anfitrion ]
        fotoPerfil:
          type: string
          format: uri
        descripcion:
          type: string
    updatePassword:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: Contraseña actual del usuario
        newPassword:
          type: string
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*$"
          description: Nueva contraseña del usuario (debe contener al menos un dígito, una letra minúscula y una letra mayúscula)
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "Contraseña actualizada exitosamente"
      required:
        - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT